# Improve memory access patterns for index operations.