@yongjik 2018-01-05
I have measured execution times for various configurations, and found that the performance sometimes improves a lot, while it almost never worsens.

A somewhat realistic example of big improvement is:

    B = torch.cuda.FloatTensor(2000, 2000)
    A = torch.cuda.FloatTensor(2000, 500)

    idxs = (np.arange(500) * 7) % 500
    idxs = torch.cuda.LongTensor(idxs)

    B.index_add_(1, idxs, A)
    # On GTX 1080, execution time decreases 88% (1.195 ms -> 0.138 ms).

One (arguably pathological) case where performance degrades is:

    B = torch.cuda.FloatTensor(2000, 4)
    idxs = torch.cuda.LongTensor(np.zeros(40))
    B.index_fill_(1, idxs, 1.0)
    # On GTX 1080, execution time increases 22% (13.45 us -> 16.41 us).