@yongjik 2018-01-05
Hi @apaszke thanks for the review.  You raise a good argument.

I have checked several hundred random(-ish) configurations, and found that the performance almost never becomes visibly worse, except for the weird case I mentioned above.  I'm currently running more tests including bigger strides; I'll update the thread when I get the result.