@apaszke 2018-01-05
Thanks! Looking forward to benchmark results!