@yongjik 2018-01-11
Hi, sorry for the delay, but I have run some experiments and here's the result: https://github.com/yongjik/pt_test/tree/master/results/index_select

Here's some highlights.

Among the 81869 configurations I ran some index operations, kernel execution time changed by:

      =========================
      BAD   > 20%        :     6
      Bad   10 .. 20%    :    15
      bad   5 .. 10%     :    34
            -5 .. 5%     : 79785
      good  -10 .. -5%   :   744
      Good  -20 .. -10%  :   482
      GOOD  -50% .. -20% :   472
      BEST  < -50%       :   331

Moreover, all six BAD cases (execution time increased by up to 21%) belong to the weird `index_fill_` I mentioned in my first comment above.  It's possible that I'm missing some cases: after all, I created all these tests by going through permutations, so there's a lot of "variations of the same theme".  However, the trend seems quite positive to me.

A realistic example is in line 84354 of https://raw.githubusercontent.com/yongjik/pt_test/master/results/index_select/SCORES.txt

(Sorry for the data dump.  I wasn't planning to show it to anyone, so the format is rather ugly.  You might want to download the file and open it in some editor.)

Here, we have source tensor of (512, 255) and dest tensor of (512, 256), applying index operations on columns.  Everything I can throw at it shows 9-15% improvement.  See the first link above for more examples.  (Some large tensors show up to 13x improvement.)