@apaszke 2018-01-11
That's great thanks for doing such a detailed benchmark! Do you know why is this case slower after this patch?