@yongjik 2018-01-11
Well, in that example `index_fill_()` is trying to write to the same column 40 times, so of course it will lead to write conflicts.  In the original code, at least these 40 writes are spread far apart (probably in different thread blocks), but in the new code they are executed by 40 consecutive kernel threads.

However, I think one can reasonably argue that such a use of `index_fill_` is "broken" performance-wise, and that we cannot expect the code to be fast anyway.