@yongjik 2018-01-16
@apaszke Hi, any more thoughts on this PR?