@soumith 2018-01-22
we should merge this in.