@soumith 2018-01-22
@pytorchbot test this please