@apaszke 2018-01-22
@yongjik sorry I didn't have time to review this again. It looks good at a glance, so we can merge it